database improvements

+++ src/config.h
@@ -42,8 +42,8 @@ namespace mc {
 #ifdef MC_KMER_TYPE
     using kmer_type = MC_KMER_TYPE ;
 #else
-    using kmer_type = std::uint32_t;
-//    using kmer_type = std::uint64_t;
+//    using kmer_type = std::uint32_t;
+    using kmer_type = std::uint64_t;
 #endif
 
 
@@ -87,10 +87,6 @@ using sequence = std::string;
  * @brief controls how nucleotide sequences are transformed into 'features'
  */
 using sketcher = single_function_min_hasher<kmer_type>;
-//using sketcher = single_function_min_hasher32;
-//using sketcher = minimizer_hasher<kmer_type>;
-//using sketcher = single_function_statistics_min_hasher;
-//using sketcher = entropy_hasher<kmer_type>;
 
 
 /**************************************************************************
